requirements.txt
torch
transformers
sentence-transformers
xgboost
scikit-learn
fastapi
uvicorn
pydantic
datasets
pytest
structlog
pyyaml
